snapshot: Add accessors for updating snapshot list relations
Rather than allowing a leaky abstraction where multiple drivers have
to open-code operations that update the relations in a
virDomainSnapshotObjList, it is better to add accessor functions so
that updates to relations are maintained closer to the internals.  The
goal here is to avoid access to, nchildren, first_child, or sibling
outside of the lowest level functions, making it easier to refactor
later on.  While many of the conversions are fairly straightforward,
the MoveChildren refactoring can be a bit interesting to follow.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320054105.17689-8-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>