snapshot: Refactor list filtering
Separate the algorithm for which list members to vist (which is
generic and can be shared with checkpoints, provided that checkpoints
pick the same bit values for some of its flags) from the decision on
which members to return (which is specific to snapshots).  The typedef
for the callback function feels a bit heavy here, but will make it
easier to move the common portions in a later patch.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320054105.17689-10-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>