snapshot: Rename file for virDomainMomentObj
Now that we have made virDomainMomentObj sufficiently generic to
support both snapshots and checkpoints, it is time to rename the file
that it lives in. The split between a generic object and a list of the
generic objects doesn't buy us as much, so it will be easier to stick
all the moment list code in one file, with more code moving in the
next patch.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320054105.17689-14-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>