snapshot: Move snapshot list code into generic file
Finish the code motion of generic moment list handling. Well, mostly -
we need to convert to using virObject to get polymorphic cleanup
functions (so for now there are still a few lingering ties specific to
snapshots).  In this case, I kept virDomainSnapshotObjList as a
wrapper type around the new generic virDomainMomentObjList; the bulk
of the algorithms move, but the old functions remain and forward to
the generic helpers.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320054105.17689-15-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>