snapshot: Tweaks to support new bulk dumpxml/import API
Change the return value of virDomainSnapshotObjLisParse() to return
the number of snapshots imported, and allow a return of 0 (the
original proposal of adding a flag to virDomainSnapshotCreateXML
required returning an arbitrary non-NULL snapshot, but with a new API
that returns a count, we are no longer constrained to a non-empty
list).

Document which flags are supported (namely, just SECURE) in
virDomainSnapshotObjListFormat().

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320054105.17689-16-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>