snapshot: Store qemu snapshot state in bulk
Rather than one file per snapshot, store all qemu snapshots in a
single file, using the recently added bulk snapshot list
operations. For now, this doesn't change how often libvirt writes a
snapshot file, but it does open the door for the next patch to update
the signature to qemuDomainSnapshotWriteMetadata() and call it less
frequently.  One of the main benefits for doing a bulk write is that
you only have to do a single file system write at the end of an
operation, rather than potentially 3 during
virDomainSnapshotCreateXML(REDEFINE|CURRENT) (delete the old snapshot
definition being redefined, rewrite the previous current snapshot to
longer be current, and store the new snapshot definition) or even more
during virDomainSnapshotDelete(DESCENDENTS) (a file system hit per
snapshot being deleted).  It also makes it perhaps a bit more feasible
to roll back to earlier state if something fails horribly midway
through an operation (until you write the new file, the old file is
still a reliable record of what state to roll back to), compared to
the current code which has to track lots of things locally; although I
did not attempt to play with any patches along those lines.

Another benefit of the bulk write - it's less code to maintain, and
will make it easier for me to model qemu's checkpoint storage in the
same way (and for checkpoints, I don't even have to worry about legacy
parsing).

This is a one-way upgrade - if you have snapshots created by an older
libvirt, the new libvirt will correctly load those snapshots and
convert to the new format. But as the new libvirt will no longer
output the old format, reverting back to the old libvirt will make it
appear that all snapshots have disappeared (merely hidden until you
upgrade libvirt again).  But then again, we've never promised that
downgrading libvirt after an upgrade was supposed to work flawlessly.

There is a slight chance for confusion if a user named two separate
domains 'foo' and 'foo.xml'; the new scheme expects 'foo.xml' to be a
regular file for the former domain, while the old scheme expected
'foo.xml' to be a directory for the latter domain; if we are worried
about that, we could tweak the code to instead output new state in a
file named 'foo' instead of the more typical 'foo.xml' and just key
off of whether the file is regular or a directory when deciding if
this is the first libvirt run after an upgrade. But I felt that the
chances of a user abusing domain names like that is not worth the
effort.

The bulk snapshot file can be huge (over RPC, we allow <domain> to be
up to 4M, and we allow up to 16k snapshots; since each each snapshot
includes a <domain>, a worst-case domain would result in gigabytes of
bulk data); it is no worse on overall filesystem usage than before,
but now in a single file vs. a series of files it requires more memory
to read in at once.  I don't know if we have to worry about that in
practice, but my patch does cap things to read in no more than an
arbitrarily-picked 128M, which we may have to raise in the future.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320213242.26859-1-eblake@redhat.com>