qemu_hotplug: pull qemuDomainUpdateDeviceList out of qemuDomainDetachDeviceLive
qemuDomainDetachDeviceLive() is called from two places in
qemu_driver.c. Move the call to qemuDomainUpdateDeviceList() from the
end of that function, which is now in qemu_hotplug.c, to the two
places that call qemuDomainDetachDeviceLive().

Removing that one call from qemuDomainDetachDeviceList() will permit
us to call it from the test driver, replacing the separate calls to
qemuDomainDetachDeviceDiskLive(), qemuDomainDetachChrDevice(),
qemuDomainDetachShmemDevice() and qemuDomainDetachWatchdog(). We want
to do this so that part of the common functionality of those three
functions (and the rest of the device-specific Detach functions) can
be pulled up into qemuDomainDetachDeviceLive() without breaking the
test. (This is done in the next patch).

Signed-off-by: Laine Stump <laine@laine.org>
Message-Id: <20190321222901.12902-16-laine@laine.org>