qemu: monitor: move event data structure to domain
In order to later process the event data, we need to be able to access
it over the lifetime of the domain. So, as an initial step, let's move
the event data structure and utility functions to qemu_domain.[ch] and
rename them appropriately to make it accessible for other users.

Reviewed-by: Boris Fiuczynski <fiuczy@linux.ibm.com>
Signed-off-by: Bjoern Walk <bwalk@linux.ibm.com>
Message-Id: <20190325080457.74609-2-bwalk@linux.ibm.com>