virsh: domstate: report detailed state if available
Add a new parameter to virsh domstate, --info, to report additional
information for the domain state, e.g. for a QEMU guest running a S390
domain:

    virsh # domstate --info guest-1
    crashed (panicked: s390: core='0' psw-mask='0x0002000180000000' \
             psw-addr='0x000000000010f146' reason='disabled-wait')

When the new API virDomainGetStateParams is not available for the server
or not supported by the hypervisor driver, fall back to the old API
using virDomainGetState function.

The --info parameter implies the --reason parameter and if additional
information is not available, the output is the same.

Reviewed-by: Boris Fiuczynski <fiuczy@linux.ibm.com>
Signed-off-by: Bjoern Walk <bwalk@linux.ibm.com>
Message-Id: <20190325080457.74609-7-bwalk@linux.ibm.com>