news: Document kernel requirements for virtual networks
After 7431b3eb9a05068e4b libvirt requires "filter", "nat" and
"mangle" tables to exist for both IPv4 and IPv6. This fact was
missed in the news.xml and since we don't have any better place
to advertise that let's update old news.

This was refined in 686803a1a2e and since that is not released
yet create a new entry documenting the refinement.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <82be5fa203ddc31b9ce683cee6e76630b575e759.1553509899.git.mprivozn@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>