backup: Introduce virDomainBackup APIs
Introduce a few more new public APIs related to incremental backups.
This builds on the previous notion of a checkpoint (without an
existing checkpoint, the new API is a full backup, differing from
virDomainBlockCopy in the point of time chosen and in operation on
multiple disks at once); and also allows creation of a new checkpoint
at the same time as starting the backup (after all, an incremental
backup is only useful if it covers the state since the previous
backup).  Snapshot creation is also a point in time at which creating
a checkpoint atomically can be useful. A backup job also affects
filtering a listing of domains, as well as adding event reporting for
signaling when a push model backup completes (where the hypervisor
creates the backup); note that the pull model does not have an event
(starting the backup lets a third party access the data, and only the
third party knows when it is finished).

Since multiple backup jobs can be run in parallel in the future (well,
qemu doesn't support it yet, but we don't want to preclude the idea),
virDomainBackupBegin() returns a job id, which can also be queried by
virDomainListJobIds(), and this job id must be used for
virDomainBackupGetXMLDesc() and virDomainBackupEnd(). In the future,
we may also extend other jobs (migration as the default global job
impacting virDomainJobStats(), and the various block jobs) to all have
ids, where the existing APIs act like thin wrappers around more
powerful APIs that support a job id everywhere.

The full list of new API:
        virDomainBackupBegin;
        virDomainBackupEnd;
        virDomainBackupGetXMLDesc;
        virDomainListJobIds;
        virDomainSnapshotCreateXML2;

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190326061353.7206-9-eblake@redhat.com>