tests: Disable some tests on 32 bit systems
On 32 bit systems there are two tests failing currently:
qemufirmwaretest and qemuxml2argvtest (not every test case in
them is failing, only some of them). There are several problems:

1) in qemufirmwaretest it's 'QEMU FW precedence test' which is
failing. This is because the code tests
qemuFirmwareFetchConfigs() which tries to compile a list of some
paths from the system and if one of the paths is an empty file it
should not appear on the returned list. The code uses
virFileLength() to query the file size which uses stat() under
the hood. The test uses virfilewrapper to redirect stat() into
qemufirmwaredata/ but for reasons beyond me real_stat() returns
mangled buffer making the code see a file with a real size even
for empty files. I've track this one to a problem in
virfilewrapper.c which will call real___xstat() which points to a
function in libvirt.so (!) which returns stat with 32bit members
even though we're compiling with LARGE_FILE and therefore expect
64bit members in the stat struct.

2) qemuxml2argvtest has two tests failing:
2a) [aarch64-]os-firmware-* - These fail because of the same
reason as described in 1)

2b) pseries-hostdevs-* - These fail because they again rely on
some stat (lstat to be precise). When building the cmd line for
these test cases the code does some search in /sys/bus/pci/...
and since we have virpcimock everything works just fine. Except
when virfilewrapper.c is linked in because then lstat() is taken
from there and when it initializes itself it will lookup
real_lstat() which will now point not to virpcimock but to glibc.
Again, reasons beyond me.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <b0c7425b477879d739fad3435cacc95af706e038.1553772560.git.mprivozn@redhat.com>